refactor: 이미지업로더 유틸 분리 및 회원가입 타입 분리
