+++ src/stores/useSignupStore.ts
@@ -1,5 +1,7 @@
 import { create } from "zustand";
 
+import { SignupFieldKey } from "@/types/signup";
+
 interface SignupState {
   name: string;
   email: string;
@@ -11,7 +13,7 @@ interface SignupState {
   region: string;
   profileimg: string;
 
-  setField: (key: keyof SignupState, value: string) => void;
+  setField: (key: SignupFieldKey, value: string) => void;
   setAgreed: (ids: number[]) => void;
   reset: () => void;
 }
@@ -27,7 +29,6 @@ export const useSignupStore = create<SignupState>(set => ({
   profileimg: "",
 
   setField: (key, value) => set(state => ({ ...state, [key]: value })),
-
   setAgreed: ids => set({ agreed: ids }),
 
   reset: () =>
