+++ src/types/signup.ts
@@ -1,0 +1,8 @@
+export type SignupFieldKey =
+  | "name"
+  | "email"
+  | "phone"
+  | "nickname"
+  | "gender"
+  | "region"
+  | "profileimg";
